bot/bot_api/requirements.txt
python-dotenv==1.0.1
python-telegram-bot==21.6
redis==5.2.0
setuptools==75.4.0
aiocache==0.12.3